Refactor new discovery code from index.ts and move it
